api/requirements.txt
fastapi>=0.115.0
uvicorn>=0.32.0
chromadb>=0.5.0
openai>=1.50.0
anthropic>=0.39.0
python-dotenv>=1.0.0
slowapi>=0.1.9
